hugo/DefinableText: Refactored DefinableText
Moved non gesture functions out of the coordinator and into a newly created DefinableTextViewModel. This fixes a bug where the Coordinator does not recieve the lastest tokensRange, which caused tap gesture to not highlight words.
